Its... done..? Optimizations are possible, and bad programming habits are evident here and there. But it IS FULLY working now. Only thing missing would be a way for a user to upload video to the program. But setting that up when this actually isn't gonna be used, wouldn't be super kewl.
Biggest changes this commit, is the addition of a way for the program to handle it finding several windows it believes could be the upload page. It handles it by just putting them in the foreground and closing them down via ctrl-w... lmao. Then normal program resumes where it opens its own new upload page.
Chaining uploads also possible now, due to the small but important addition of the program closing the upload tab when its fully complete and uploaded.

Main thing that can break it right now, would be multi monitor setups, and chrome randomly deciding to open somewhere else :cowboi: